requirements.txt
python-dotenv
openai
IPython